Enemigos persiguen al jugador más cercano

+++ Save_the_world/enemy.h
@@ -8,7 +8,7 @@
 
 class enemy: public QObject, public QGraphicsItem
 {
-    int posx ,posy;
+    float posx ,posy;
     int velocidad=2;
     int r=9;
 
@@ -23,10 +23,10 @@ public:
     void left();
     void right();
 
-    void setPosx(int value);
-    void setPosy(int value);
-    int getPosx() const;
-    int getPosy() const;
+    void setPosx(float value);
+    void setPosy(float value);
+    float getPosx() const;
+    float getPosy() const;
     int getR() const;
 
     int getVelocidad() const;
